// Views/src/components/financial/components/cash-withdrawn-expense/dto/list-cash-withdrawn-expenses-dto.ts
import { ItemsInterface } from "src/shared/components/list-g/list/interfaces/items-interface";



export class ListCashWithdrawnExpensesDto {
  id: ItemsInterface;
  name: ItemsInterface;
  place:ItemsInterface;
  category: ItemsInterface;
  subcategory: ItemsInterface;
  withdrawnOn: ItemsInterface;
  withdrawnOnToView: ItemsInterface;
  price: ItemsInterface;
}